==> 01-notebooks/01-import.ipynb <==
import json
import requests
import praw
import os

reddit = praw.Reddit(
    client_id = "secret",
    client_secret="secret",
    user_agent="export wiki pages/1.0 "
)

wikis =  reddit.subreddit("germany").wiki

keep_list = ["american-dream",
"autobahn_safety",
"blue-card",
"citizenship",
"citizenship-detour",
"culture/etiquette",
"culture/food",
"gun-deaths",
"health_insurance",
"how-to-germany",
"how-to-study",
"immigrants",
"liability-insurance",
"living",
"living/anmeldung",
"living/banking",
"living/buying-property",
"living/cannabis",
"living/car",
"living/climate",
"living/connectivity",
"living/downloads",
"living/finances",
"studying/housing",
"living/knowing-german",
"living/problems",
"living/rundfunkbeitrag",
"living/savingmoney",
"living/shipping",
"living/shoping",
"living/why",
"overton-window",
"paths",
"progressive",
"studying",
"studying/financing",
"studying/general",
"studying/guide",
"studying/immigration",
"studying/research",
"studying/resources",
"taxes",
"travelling",
"working/findingajob",
"working/freelance",
"working/parental_leave",
"working/remote",
"working/requirements"]

wiki_dict = {}

for page in keep_list:
    content = wikis[page].content_md
    wiki_dict[page] = content

folder = "./wikis"

p = "https://www.reddit.com/r/germany/wiki"

for name, content in wiki_dict.items():
    new_name = name.replace("/", "-")
    with open(f"{folder}/{new_name}.md", "w", encoding="utf-8") as file:
        file.write(f"{p}/{name}\n\n")
        file.write(content)
