==> 01-notebooks/03-chunk.ipynb <==
from openai import OpenAI
import os
import pandas as pd
from tqdm.auto import tqdm
import json
import uuid

client = OpenAI(api_key=os.environ["IA"])

test = pd.read_json("./documents.json").to_dict(orient="records")

chunk_list = []
ids = []

for i in test:
    if i["length"] > 1700:
        chunk_list.append({"source": i["source"], "content": i["content"], "headline":i["headline"], "length":i["length"]})
        ids.append(i["id"])

len(ids), ids[0]

filtered_list = [record for record in test if record['id'] not in ids]

len(filtered_list) + len(chunk_list) == len(test)

def fetch_knowledge_chunks(Chunk):
    content = Chunk["content"]
    source = Chunk["source"]
    headline = Chunk["headline"]
    prompt = f"""Task: Extract Knowledge Chunks
    Please extract knowledge chunks from the following content.
    Return the extracted knowledge chunks as a JSON object or array (even if there is a 1 chunk), 
    ensuring that each chunk approximately 1000 character (not strict rule).
    If text is approximately 2000 character extract more than 1 chunk.
    Text is includes markdown format. keep link, list, etc. Use the format: 
    {{"chunks": "chunk"}}
    
    
    Text:

    {content}
    """

    completion = client.chat.completions.create(
        model="gpt-4o-mini",
        response_format={"type": "json_object"},
        messages=[
            {
                "role": "system",
                "content": "You are an assistant that takes apart a piece of text into semantic chunks to be used in a RAG system.",
            },
            {"role": "user", "content": prompt},
        ],
        stream=False,
    )

    res = json.loads(completion.choices[0].message.content)

    return res, source, headline

len(chunk_list[0]["content"])

chundked_list=[]

for chunk in tqdm(chunk_list):
    res, source, headline = fetch_knowledge_chunks(chunk)
    for r in res["chunks"]:
        if len(r)==1:
            r = r["chunk"] or r["text"]
            chundked_list.append({"source":source, "headline":headline, "content":r, "length": len(r), "id": str(uuid.uuid4())})
        else:
            chundked_list.append({"source":source, "headline":headline, "content":r, "length": len(r), "id": str(uuid.uuid4())})

len(chundked_list)

filtered_list[1]

chundked_list[1]

doc_chunked = filtered_list + chundked_list
len(doc_chunked)

with open('doc_chunked.json', 'wt') as f_out:
    json.dump(doc_chunked, f_out, indent=2)
