==> 01-notebooks/04-get_question.ipynb <==
## Chatgpt4o-mini

from openai import OpenAI, Client
import os
import pandas as pd
from tqdm.auto import tqdm
import json
from IPython.display import display, clear_output, Markdown


with open("doc_chunked.json", 'rt') as f_in:
    documents = json.load(f_in)

len(documents)

documents[0]

client = OpenAI(api_key=os.environ["IA"])

prompt_template = """
You're a helpful assistant. For each content provided, generate one  question that a user might ask. The question should be directly relevant and focus on the key message or detail within the content. If possible, use as fewer words as possible from the content.The questions should be complete and not too short.


HEADLINE: {headline}
CONTENT: {content}

return just question as string.

""".strip()

def get_prompt(doc):
    content = doc["content"]
    headline = doc["headline"]
    promt = prompt_template.format(content=content, headline=headline)
    return promt

def gen_q(prompt):
    response = client.chat.completions.create(
        model='gpt-4o-mini',
        messages=[{"role": "user", "content": prompt}]
    )

    res = response.choices[0].message.content
    return res

print(get_prompt(documents[1]))

documents_new= []

for doc in tqdm(documents):
    promt = get_prompt(doc)
    res = gen_q(promt)
    id = doc["id"]
    documents_new.append({"id":id, "question":res})

documents_new[0]

question_map = {q['id']: q['question'] for q in  documents_new}
test = list(map(lambda doc: {**doc, 'question': question_map.get(doc['id'], None)}, documents))
len(test)

test[0]

with open('doc_with_q_4o-mini.json', 'wt') as f_out:
    json.dump(test, f_out, indent=2)

## Llama 3.1

llama3 = Client(
    base_url='http://localhost:11434/v1/',
    api_key='ollama',
)

def generate_questions(prompt):
    response = llama3.chat.completions.create(
        model="llama3.1:70b",
        messages=[{"role": "user", "content": prompt}]
    )

    json_response = response.choices[0].message.content
    return json_response

generate_questions(documents[1])

documents_new_llama = []

documents[0]

question_map = {q['id']: q['question'] for q in  documents_new}
test1 = list(map(lambda doc: {**doc, 'question': question_map.get(doc['id'], None)}, documents))
len(test1)

for doc in tqdm(documents):
    promt = get_prompt(doc)
    res = generate_questions(promt)
    id = doc["id"]
    documents_new_llama.append({"id":id, "question":res})
    

with open('doc_with_q_lama.json', 'wt') as f_out:
    json.dump(test1, f_out, indent=2)

test1[0]
