==> 01-notebooks/05-get_grand_truth.ipynb <==
import os 
from openai import Client
import json
import pandas as pd

with open("docs_with_q_4o-mini.json", 'rt') as f_in:
    list = json.load(f_in)

len(list)

client = Client(api_key=os.environ["IA"])

list[1]

prompt_template = """
You're a RAG project assistant. Generate 5 question user might ask based on the FAQ record. The record should contain the answer to the questions, and the questions should be complete and not too short.
If possible, use as fewer words as possible from the record. 

The record:

Headline: {headline}
Question: {question}
Content: {content}


Provide the output in parsable JSON without using code blocks:

["question1", "question2", ..., "question5"]
""".strip()

def generate_questions(doc):
    prompt = prompt_template.format(**doc)

    response = client.chat.completions.create(
        model='gpt-4o-mini',
        messages=[{"role": "user", "content": prompt}]
    )

    json_response = response.choices[0].message.content
    return json_response

test = generate_questions(list[1])

from tqdm.auto import tqdm

json.loads(test)

results = {}

for doc in tqdm(list):
    doc_id = doc["id"]
    questions = generate_questions(doc)
    results[doc_id] = questions


len(results)

index =  {d["id"]: d for d in list}

final_q = []

parsed_resulst = {}

for doc_id, json_questions in results.items():
    parsed_resulst[doc_id] = json.loads(json_questions)


for doc_id, questions in parsed_resulst.items():
    headline = index[doc_id]["headline"]
    for q in questions:
        final_q.append((q, headline, doc_id))

final_q[2]

df = pd.DataFrame(final_q, columns=['question', 'headline', 'content'])
len(df)

df.to_csv("./gp4o-mini-questions.csv", index=False)
