==> 01-notebooks/06-haystack_retrieval_eval.ipynb <==
from haystack import Document
import json
import pandas as pd

with open("docs_with_q_4o-mini.json", "rt") as f_in:
    ds_gpt = json.load(f_in)

ds_gpt[0]

ds_gpt[660]

ds_gpt[47]

with open("docs_with_q_lama.json", "rt") as f_in:
    ds_llama = json.load(f_in)

part2 = ds_llama[:47] 

part1 = ds_llama[47:661]

ds_llama = part1 + part2

ds_llama[0]

ds_llama[660]

ds_llama[47]


def prepare_docs(prepared_docs, docs):
    for ds in docs:
        prepared_docs.append(
            Document(content=ds["content"], meta={"source": ds["source"], "headline": ds["headline"], "ids": ds["id"], "length":ds["length"], "question": ds["question"]})
        )
    return prepared_docs

from haystack_integrations.document_stores.elasticsearch import ElasticsearchDocumentStore
from haystack.components.embedders import SentenceTransformersTextEmbedder, SentenceTransformersDocumentEmbedder
from haystack.components.writers import DocumentWriter
from haystack.document_stores.types import DuplicatePolicy
from haystack import Pipeline
from haystack_integrations.components.retrievers.elasticsearch import ElasticsearchEmbeddingRetriever

def prepare_document_store(model, docs, index_name):
    document_store = ElasticsearchDocumentStore(hosts = "http://localhost:9200", embedding_similarity_function="cosine", index=index_name)
    document_embedder = SentenceTransformersDocumentEmbedder(model=model,
                                                         meta_fields_to_embed=["headline", "question"])
    document_writer = DocumentWriter(document_store=document_store, policy=DuplicatePolicy.SKIP)
    indexing = Pipeline()
    indexing.add_component(instance=document_embedder, name="document_embedder")
    indexing.add_component(instance=document_writer, name="document_writer")
    indexing.connect("document_embedder.documents", "document_writer.documents")
    indexing.run({"document_embedder": {"documents": docs}}),
    return document_store


def prepare_retrieval_pipeline(model, document_store):
    retrieval_pipeline = Pipeline()
    retrieval_pipeline.add_component("text_embedder", SentenceTransformersTextEmbedder(model=model, progress_bar=False))
    retrieval_pipeline.add_component("retriever", ElasticsearchEmbeddingRetriever(document_store=document_store, top_k=5))
    retrieval_pipeline.connect("text_embedder", "retriever")
    return retrieval_pipeline

import pandas as pd

df_ground_truth = pd.read_csv("gp4o-mini-questions.csv")
ground_truth = df_ground_truth.to_dict(orient="records")


ground_truth[0]

def prepare_ground_truth_docs(docs):
    ground_truth_docs = []
    for doc in docs:
        ground_truth_docs.append(
            [Document(content=doc["content"]), Document(content=doc["content"]),Document(content=doc["content"]),Document(content=doc["content"]),Document(content=doc["content"])])
        ground_truth_docs.append(
            [Document(content=doc["content"]), Document(content=doc["content"]),Document(content=doc["content"]),Document(content=doc["content"]),Document(content=doc["content"])])
        ground_truth_docs.append(
            [Document(content=doc["content"]), Document(content=doc["content"]),Document(content=doc["content"]),Document(content=doc["content"]),Document(content=doc["content"])])
        ground_truth_docs.append(
            [Document(content=doc["content"]), Document(content=doc["content"]),Document(content=doc["content"]),Document(content=doc["content"]),Document(content=doc["content"])])
        ground_truth_docs.append(
            [Document(content=doc["content"]), Document(content=doc["content"]),Document(content=doc["content"]),Document(content=doc["content"]),Document(content=doc["content"])])
        
    return ground_truth_docs

from tqdm.auto import tqdm

def prepare_retrieved_docs(retrieval_pipeline):
    retrieved_docs = []
    for q in tqdm(ground_truth):
        question = q["question"]
        result = retrieval_pipeline.run({"text_embedder": {"text": question}})
        retrieved_docs.append([result["retriever"]["documents"][0], result["retriever"]["documents"][1], result["retriever"]["documents"][2] ,result["retriever"]["documents"][3], result["retriever"]["documents"][4]])
    return retrieved_docs

from haystack.components.evaluators import DocumentMRREvaluator, DocumentRecallEvaluator, DocumentMAPEvaluator

def prepare_eval():
    eval_pipe = Pipeline()
    eval_pipe.add_component("doc_mrr_evaluator", DocumentMRREvaluator())
    eval_pipe.add_component("doc_rec_evaluator", DocumentRecallEvaluator())
    eval_pipe.add_component("map", DocumentMAPEvaluator())
    return eval_pipe

def run(data, model, index_name):
    prepared_docs = []
    docs = prepare_docs(prepared_docs, data)
    document_store = prepare_document_store(model=model, docs=docs, index_name=index_name)
    retrieval_pipeline = prepare_retrieval_pipeline(model=model, document_store=document_store)
    ground_truth_docs = prepare_ground_truth_docs(ds_gpt)
    retrieved_docs = prepare_retrieved_docs(retrieval_pipeline)
    eval_pipe = prepare_eval()
    results = eval_pipe.run(
        {
            "doc_mrr_evaluator": {
                "ground_truth_documents": ground_truth_docs,
                "retrieved_documents": retrieved_docs,
            },
            "doc_rec_evaluator": {
                "ground_truth_documents": ground_truth_docs,
                "retrieved_documents": retrieved_docs,
            },
            "map": {
                "ground_truth_documents": ground_truth_docs,
                "retrieved_documents": retrieved_docs,
            },
        }
    )
    return results



from itertools import product


run_dicts = {
    "models": {
        "model_mini_lm": "sentence-transformers/multi-qa-MiniLM-L6-cos-v1",
        "model_distilbert": "sentence-transformers/multi-qa-distilbert-cos-v1"

    },
    "datasets": {
        "docs_4o_mini": ds_gpt,
        "docs_llama" : ds_llama
    }
}

models = list(run_dicts["models"].items())
datasets = list(run_dicts["datasets"].items())

for (model_key, model_value), (dataset_key, dataset_value) in product(models, datasets):
    print(f"Model Key: {model_key}, Model Value: {model_value}, Dataset: {dataset_key}")
    index_name = f"{model_key}_{dataset_key}"
    print(index_name)

results_dict = {}

for (model_key, model_value), (dataset_key, dataset_value) in product(models, datasets):
    model = model_value
    dataset = dataset_value
    dataset_name = dataset_key
    index_name = f"{model_key}_{dataset_key}"
    results = run(data=dataset, model=model, index_name=index_name)

    results_dict[index_name]= {}
    for evaluator in results:
        results_dict[index_name][evaluator] = results[evaluator]["score"]

results_dict

pd.DataFrame.from_dict(results_dict, orient="index")
