==> 01-notebooks/07-retrieval_eval.ipynb <==
import json
from elasticsearch import Elasticsearch, helpers


with open("docs_with_q_4o-mini.json", "rt") as f_in:
    docs_4o_mini = json.load(f_in)

with open("docs_with_q_lama.json", "rt") as f_in:
    docs_llama = json.load(f_in)

from tqdm.auto import tqdm

from sentence_transformers import SentenceTransformer

model_mini_lm = SentenceTransformer("multi-qa-MiniLM-L6-cos-v1")
model_distilbert = SentenceTransformer("multi-qa-distilbert-cos-v1")

lenv_mini = len(model_mini_lm.encode("test"))
lenv_distilbert = len(model_distilbert.encode("test"))

es_client = Elasticsearch("http://localhost:9200")

docs_4o_mini[0]

def index_sett(v):

    index_settings = {
        "settings": {
            "number_of_shards": 1,
            "number_of_replicas": 0
        },
        "mappings": {
            "properties": {
                "source": {"type": "text"},
                "content": {"type": "text"},
                "headline": {"type": "text"},
                "id": {"type": "keyword"},
                "length": {"type": "integer"},
                "question": {"type": "text"},
                "question_vector": {
                    "type": "dense_vector",
                    "dims": v,
                    "index": True,
                    "similarity": "cosine"
                },
                "content_vector": {
                    "type": "dense_vector",
                    "dims": v,
                    "index": True,
                    "similarity": "cosine"
                },
                "question_content_vector": {
                    "type": "dense_vector",
                    "dims": v,
                    "index": True,
                    "similarity": "cosine"
                },
            }
        }
    }

    return index_settings

from itertools import product



emb = {
    "model": {
        model_mini_lm:lenv_mini,
        model_distilbert :lenv_distilbert,

    },
    "dataset": {
        "docs_4o_mini": docs_4o_mini,
        "docs_llama" : docs_llama
    }
}

datasets = emb["dataset"]
"384_docs_4o_mini"

emb["dataset"].keys()

models = list(emb["model"].items())
datasets = list(emb["dataset"].items())

for (model_key, model_value), (dataset_key, dataset_value) in product(models, datasets):
    print(f"Model Key: {model_key}, Model Value: {model_value}, Dataset: {dataset_key}")



for (model_key, model_value), (dataset_key, dataset_value) in product(models, datasets):

    index_name = f"{str(model_value)}_{dataset_key}"
    settings = index_sett(model_value)
    print(index_name, model_value)
    es_client.indices.delete(index=index_name, ignore_unavailable=True)
    es_client.indices.create(index=index_name, body=settings)

from tqdm.auto import tqdm

docs_4o_mini[0]

def import_doc(model_key, dataset):
    for doc in tqdm(dataset):
        question = doc["question"]
        content = doc["content"]
        qt = question + ' ' + content
        doc["question_vector"] = model_key.encode(question)
        doc["content_vector"] = model_key.encode(content)
        doc["question_content_vector"] = model_key.encode(qt)
    return

print(docs_llama[395]["content"])

dataset_key

def generate_actions(index_name, data):
    for i, record in enumerate(data):
        yield {
            "_index": index_name,
            "_id": i,
            "_source": record,
        }


from tqdm.auto import tqdm

import_doc(model_mini_lm, docs_4o_mini)
actions = generate_actions(index_name="384_docs_4o_mini", data=docs_4o_mini)
helpers.bulk(es_client, actions)


import_doc(model_mini_lm, docs_llama)
actions = generate_actions(index_name="384_docs_llama", data=docs_llama)
helpers.bulk(es_client, actions)


import_doc(model_distilbert, docs_4o_mini)
actions = generate_actions(index_name="768_docs_4o_mini", data=docs_4o_mini)
helpers.bulk(es_client, actions)

import_doc(model_distilbert, docs_llama)
actions = generate_actions(index_name="768_docs_llama", data=docs_llama)
helpers.bulk(es_client, actions)

import pandas as pd

df_ground_truth = pd.read_csv("gp4o-mini-questions.csv")
ground_truth = df_ground_truth.to_dict(orient='records')
ground_truth[0]

def elastic_search_knn(field, vector, index_name):
    knn = {
        "field": field,
        "query_vector": vector,
        "k": 5,
        "num_candidates": 10000,
        }

    search_query = {
        "knn": knn,
        "_source": ["content", "headline", "question", "source", "length", "id"]
    }

    es_results = es_client.search(
        index=index_name,
        body=search_query
    )
    
    result_docs = []
    
    for hit in es_results['hits']['hits']:
        result_docs.append(hit['_source'])

    return result_docs



def hit_rate(relevance_total):
    cnt = 0

    for line in relevance_total:
        if True in line:
            cnt = cnt + 1

    return cnt / len(relevance_total)


def mrr(relevance_total):
    total_score = 0.0

    for line in relevance_total:
        for rank in range(len(line)):
            if line[rank] == True:
                total_score = total_score + 1 / (rank + 1)

    return total_score / len(relevance_total)


def evaluate(ground_truth, search_function, model, index_name):
    relevance_total = []

    for q in tqdm(ground_truth):
        doc_id = q["content"]
        results = search_function(q, model, index_name)
        relevance = [d['id'] == doc_id for d in results]
        relevance_total.append(relevance)

    return {
        'hit_rate': hit_rate(relevance_total),
        'mrr': mrr(relevance_total),
    }

def question_vector_knn(q, model, index_name):
    question = q['question']

    v_q = model.encode(question)

    return elastic_search_knn('question_vector', v_q, index_name)

for (model_key, model_value), (dataset_key, dataset_value) in product(models, datasets):
    index_name = f"{model_value}_{dataset_key}"
    print(index_name, model_value)
    print(evaluate(ground_truth, question_vector_knn, model_key, index_name), index_name)
    

def content_vector_knn(q, model, index_name):
    question = q['question']

    v_q = model.encode(question)

    return elastic_search_knn("content_vector", v_q, index_name)

for (model_key, model_value), (dataset_key, dataset_value) in product(models, datasets):
    index_name = f"{model_value}_{dataset_key}"
    print(evaluate(ground_truth, content_vector_knn, model_key, index_name), index_name)
    

def question_text_vector_knn(q, model, index_name):
    question = q['question']

    v_q = model.encode(question)

    return elastic_search_knn("question_content_vector", v_q, index_name)



for (model_key, model_value), (dataset_key, dataset_value) in product(models, datasets):
    index_name = f"{model_value}_{dataset_key}"
    print(evaluate(ground_truth, question_text_vector_knn, model_key, index_name), index_name)
