==> 01-notebooks/08-haystack_rag_eval.ipynb <==
from haystack import Document
import json
import pandas as pd
import os

from os import environ
import google.auth
from google.oauth2 import service_account
from google.auth.transport.requests import AuthorizedSession


SERVICE_ACCOUNT_FILE = "./creds.json"


SEARCH = os.environ["SEARCH"]

def get_token():
    credentials = service_account.IDTokenCredentials.from_service_account_file(
        SERVICE_ACCOUNT_FILE, target_audience=SEARCH)
    request = google.auth.transport.requests.Request()
    credentials.refresh(request)
    token = credentials.token

    return token

a = get_token()

SEARCH = os.environ["SEARCH"]

hed = get_token()

a = hed["Authorization"]

with open("docs_with_q_4o-mini.json", "rt") as f_in:
    ds_gpt = json.load(f_in)

def prepare_docs(prepared_docs, docs):
    for ds in docs:
        prepared_docs.append(
            Document(content=ds["content"], meta={"source": ds["source"], "headline": ds["headline"], "ids": ds["id"], "length":ds["length"], "question": ds["question"]})
        )
    return prepared_docs

from haystack_integrations.document_stores.elasticsearch import ElasticsearchDocumentStore
from haystack.components.embedders import SentenceTransformersTextEmbedder, SentenceTransformersDocumentEmbedder
from haystack.components.writers import DocumentWriter
from haystack.document_stores.types import DuplicatePolicy
from haystack import Pipeline
from haystack_integrations.components.retrievers.elasticsearch import ElasticsearchEmbeddingRetriever

def prepare_document_store(model, docs, index_name):
    document_store = ElasticsearchDocumentStore(hosts = SEARCH, embedding_similarity_function="cosine", index=index_name,request_timeout=3000,    bearer_auth=a)
    document_embedder = SentenceTransformersDocumentEmbedder(model=model,
                                                         meta_fields_to_embed=["headline", "question"])
    document_writer = DocumentWriter(document_store=document_store, policy=DuplicatePolicy.OVERWRITE)
    indexing = Pipeline()
    indexing.add_component(instance=document_embedder, name="document_embedder")
    indexing.add_component(instance=document_writer, name="document_writer")
    indexing.connect("document_embedder.documents", "document_writer.documents")
    indexing.run({"document_embedder": {"documents": docs}}),
    return document_store

def prepare_retrieval_pipeline(model, document_store):
    retrieval_pipeline = Pipeline()
    retrieval_pipeline.add_component("text_embedder", SentenceTransformersTextEmbedder(model=model, progress_bar=False))
    retrieval_pipeline.add_component("retriever", ElasticsearchEmbeddingRetriever(document_store=document_store, top_k=5))
    retrieval_pipeline.connect("text_embedder", "retriever")
    return retrieval_pipeline

prepared_docs = []

docs = prepare_docs(prepared_docs, ds_gpt)

import pandas as pd


document_store = prepare_document_store(model="./transformers", docs=docs, index_name="germany")

from elasticsearch import Elasticsearch, helpers


test = Elasticsearch()

from haystack.components.builders import PromptBuilder
from haystack.components.generators import OpenAIGenerator

import os

from getpass import getpass

os.environ["OPENAI_API_KEY"]

template = """
You're a FAQ database assistant. Answer the QUESTION based on the CONTEXT from the FAQ database.
Use only the facts from the CONTEXT when answering the QUESTION.

Context:
{% for document in documents %}
    {{ document.content }}
{% endfor %}

Question: {{question}}
Answer:
"""

prompt_builder = PromptBuilder(template=template)

generator = OpenAIGenerator(model="gpt-4o-mini")

from haystack.components.builders.answer_builder import AnswerBuilder

prepared_docs = []

docs = prepare_docs(prepared_docs, ds_gpt)

rag_pipeline = None


rag_pipeline = Pipeline()
# Add components to your pipeline



rag_pipeline.add_component("text_embedder", SentenceTransformersTextEmbedder(model="sentence-transformers/multi-qa-distilbert-cos-v1", progress_bar=False))
rag_pipeline.add_component("retriever", ElasticsearchEmbeddingRetriever(document_store=document_store, top_k=5))
rag_pipeline.add_component("prompt_builder", prompt_builder)
rag_pipeline.add_component("llm", generator)
rag_pipeline.add_component(instance=AnswerBuilder(), name="answer_builder")





rag_pipeline.connect("text_embedder", "retriever")
rag_pipeline.connect("retriever", "prompt_builder.documents")
rag_pipeline.connect("prompt_builder", "llm")
rag_pipeline.connect("llm.replies", "answer_builder.replies")
rag_pipeline.connect("llm.meta", "answer_builder.meta")
rag_pipeline.connect("retriever", "answer_builder.documents")

import pandas as pd

df_ground_truth = pd.read_csv("gp4o-mini-questions.csv")
ground_truth = df_ground_truth.to_dict(orient="records")


ground_truth[0]

ds_gpt[0]

aydi =ds_gpt[0]["id"]

combined_list = list(map(lambda gt: {**gt, **next((ds for ds in ds_gpt if ds['id'] == gt['content']), {

    "test" : gt["content"]
})}, ground_truth))

combined_list = [
    {**gt, **next((ds for ds in ds_gpt if ds['id'] == gt['content']), {})}
    for gt in ground_truth
]

new_ds = [
    {
        'question': gt['question'],
        'id': gt['content'],  
        'content': ds['content']
    }
    for gt in ground_truth
    for ds in ds_gpt
    if gt['content'] == ds['id']  
]


for gt in ground_truth[:6]:
    print(gt["question"])
    print(gt["content"])

new_ds[:6]

from tqdm.auto import tqdm

question = "test"

response = rag_pipeline.run(
{
    "text_embedder": {"text" : question},
    "prompt_builder": {"question": question},
    "answer_builder": {"query": question},
}
)
response["answer_builder"]["answers"][0].data


contexts = []
responses = []
questions = []



for run in tqdm(new_ds):
    question = run["question"]
    context = run["content"]

    questions.append(question)
    contexts.append([context])

    response = rag_pipeline.run(
    {
        "text_embedder": {"text" : question},
        "prompt_builder": {"question": question},
        "answer_builder": {"query": question},
    }
    )
    responses.append(response["answer_builder"]["answers"][0].data)
    

from haystack import Pipeline
from haystack_integrations.components.evaluators.ragas import RagasEvaluator, RagasMetric

context_utilization_pipeline = Pipeline()
evaluator = RagasEvaluator(metric=RagasMetric.CONTEXT_UTILIZATION)
context_utilization_pipeline.add_component("evaluator", evaluator)

test["contenxt"] = contexts
test["responses"] = responses
test["questions"] = questions

test.to_csv("./eval_docs.csv", header=True, sep=";")

test = pd.read_csv("./eval_docs.csv", sep=";")

t = test.sample(50)


con = t["contenxt"].to_list()

contexts = []

for c in con:
    contexts.append(list(c))

t.head()

responses = t["responses"].to_list()

questions = t["questions"].to_list()

import random

from haystack import Pipeline
from haystack_integrations.components.evaluators.ragas import RagasEvaluator, RagasMetric

answer_relevancy_pipeline = Pipeline()
evaluator = RagasEvaluator(
    metric=RagasMetric.ANSWER_RELEVANCY,
    metric_params={
        "strictness": 2
    })
answer_relevancy_pipeline.add_component("evaluator", evaluator)


evaluation_results = answer_relevancy_pipeline.run(
    {"evaluator": {"questions": questions, "contexts":  contexts, "responses": responses}}
)


sums = []

for score in (evaluation_results["evaluator"]["results"]):
    sums.append(score[0]["score"])


pd.DataFrame(sums).describe()

pd.DataFrame(sums).describe()

pd.DataFrame(sums).describe()

pd.DataFrame(sums).describe()
